LPS-65685 When the order of Upgrade classes matter, make it clear by breaking them up with a minor version increment
