+++ modules/apps/web-experience/journal/journal-service/src/main/java/com/liferay/journal/upgrade/v0_0_5/UpgradeJournalArticles.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.journal.upgrade.v1_0_0;
+package com.liferay.journal.upgrade.v0_0_5;
 
 import com.liferay.asset.kernel.model.AssetCategory;
 import com.liferay.asset.kernel.service.AssetCategoryLocalService;
