+++ modules/apps/web-experience/journal/journal-service/src/main/java/com/liferay/journal/upgrade/v0_0_5/UpgradeJournalDisplayPreferences.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.journal.upgrade.v1_0_0;
+package com.liferay.journal.upgrade.v0_0_5;
 
 import com.liferay.portal.kernel.portlet.PortletPreferencesFactoryUtil;
 import com.liferay.portal.kernel.upgrade.BaseUpgradePortletPreferences;
